Revert "RDS: RDMA: Fix the composite message user notification"
This reverts commit 941f8d55f6d613a460a5e080d25a38509f45eb75.

Subject: rds-stress stops sending on EBADSLT

This issue was caused by the following commit:

commit 941f8d55f6d6 ("RDS: RDMA: Fix the composite message user notification")

In this commit, while parsing cmsg RDS_CMSG_RDMA_ARGS, a check
was added to see if the same (sendmsg) message also contains a payload.

If so, "rdma.op_notify" is cleared and "data.op_notify" is set instead.

The objective appears to have been to queue a notification upon entering the
completion handler "rds_rdma_send_complete",
but to not queue a notification from within
"rds_send_remove_from_sock", as the later does not check for
"data.op_notify".

This obviously makes the assumption that "rds_send_remove_from_sock" is
always called prior to "rds_rdma_send_complete", which it isn't.

In the case the order expected by the author is reversed, RDS_MSG_ON_SOCK
already gets cleared from within "rds_send_remove_from_sock", and no delivery
notification is sent, since "rds_rdma_send_complete" checks for that flag.

No matter how tempting it may be to just fix this single race-condition,
the queuing of a notification from within "rds_send_remove_from_sock" must
have been done for a reason.

If it were safe to simply assume the notification will always be queued via
"rds_rdma_send_complete", this check and queuing of notifications from within
"rds_send_remove_from_sock" need not be done to begin with.

The original commit calls out for *only* notifying the application when
the full message gets delivered.

That raises a number of questions:
 1) Backwards compatibility
    How will applications behave that used to get two notifications
    (successes in the normal operations) for the two components
    (RDMA + non-RDMA) if one of those is suppressed?

 2) Partial success
    Again, how will applications behave that used to handle the
    partial success-case, but now no longer are notified about
    the difference between the entire message delivery having failed
    and the message delivery succeding only partially?

Those two questions need to be answered properly in order to
assess the impact such a change has.

In light of that, and the fact this commit does not work, it appears
the safest course of action is to just back the original commit out for now.

Orabug: 28926059

Change-Id: Ic4399e75de03101b8f8caa2bcb7c8bb4a11af51c
Signed-off-by: Gerd Rausch <gerd.rausch@oracle.com>