net/rds: Never drop the head of "clean_nodes"
Back when the RDS "xlist.h" functions were replaced by llist in
commit 1bc144b62524 ("net, rds, Replace xlist in net/rds/xlist.h with llist")

function "refill_local" (that was responsible to pop off an item
from the head of the xlist) was removed and replaced
by a much simpler way of using the head for *ibmr_ret,
and appending the rest (i.e. "clean_nodes->next") to the
"pool->clean_list".

However, ibmr_ret is _not_ always != NULL, and if it is == NULL,
we end up discarding the head of the "clean_nodes".

Apparently there was some awareness of the risk of this change,
as the original commit message of 1bc144b62524 stated:
"Known issues: don't know how to test this, need special hardware?"

Orabug: 29139555

Change-Id: I3e9927d4c8ae91eaae9dd2ca309efa4d240a00db
Signed-off-by: Gerd Rausch <gerd.rausch@oracle.com>